+++ CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Changed
+
+- Core: Turn variation parent mode back to normal if it has no non-deleted children
+
 ### Fixed
 
 - Do not rely on STATIC_URL and MEDIA_URL while formatting static and media urls
+
 
 ## [2.3.14] - 2021-02-04